refactor(ui-rnative): re-depth visualization imports after dissolving Components/
Step 3 of the DLS-1022 fold, React Native side. 80 specifiers rewritten: the 77
that reach the shared config.ts and utils/ each lose one `../`, and the 3 under
utils/ that point back into the chart components lose the now-nonexistent
Components/ path segment. Both directions are handled here, unlike the ui-react
pass where the inward ones needed a follow-up.

The rewrite is resolution-aware rather than textual. That matters most on this
side: the three `'../utils'` specifiers under CartesianChart/RevealAnimation/
resolve to the component-local CartesianChart/utils.ts, not the shared module,
and a blind sed would have silently repointed them at the wrong file -- one of
the few edits in this migration that tsc would NOT catch. They are left
untouched, along with 252 other already-correct specifiers.
