+++ libs/ui-rnative/src/lib/Components/visualization/DonutChart/DonutChart.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo } from 'react';
 import { StyleSheet, View } from 'react-native';
 import { Gesture, GestureDetector } from 'react-native-gesture-handler';
 import { scheduleOnRN } from 'react-native-worklets';
-import { chartConfig, DONUT_GEOMETRY } from '../../config';
+import { chartConfig, DONUT_GEOMETRY } from '../config';
 import { toRingLocalPoint } from './constants';
 import { DonutChartAnimatedCenter } from './DonutChartAnimatedCenter';
 import { DonutRing } from './DonutRing';
