+++ libs/ui-rnative/src/lib/Components/visualization/Line/Line.tsx
@@ -1,8 +1,8 @@
 import { useId, useMemo, memo } from 'react';
 import { Defs, G, LinearGradient, Path, Stop } from 'react-native-svg';
 
-import { chartConfig, useChartTokens } from '../../config';
-import { isNumericScale } from '../../utils/scales/scales';
+import { chartConfig, useChartTokens } from '../config';
+import { isNumericScale } from '../utils/scales/scales';
 import { useCartesianChartContext } from '../CartesianChart/context';
 import { usePathReveal } from '../CartesianChart/RevealAnimation';
