+++ libs/ui-rnative/src/lib/Components/visualization/Line/utils.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from '@jest/globals';
 import {
   getCategoricalScale,
   getNumericScale,
-} from '../../utils/scales/scales';
+} from '../utils/scales/scales';
 
 import { buildLinePath, toScaledPoints } from './utils';
