+++ libs/ui-rnative/src/lib/Components/visualization/LineChart/LineChartEmptyState/LineChartEmptyState.tsx
@@ -1,7 +1,7 @@
 import Animated from 'react-native-reanimated';
 import { G, Line as SvgLine, Path } from 'react-native-svg';
 
-import { chartConfig, useChartTokens } from '../../../config';
+import { chartConfig, useChartTokens } from '../../config';
 import { useCartesianChartContext } from '../../CartesianChart/context';
 import { useShimmerAnimation } from '../../CartesianChart/hooks/useShimmerAnimation';
 import { usePathReveal } from '../../CartesianChart/RevealAnimation';
