+++ libs/ui-rnative/src/lib/Components/visualization/LineChart/utils.ts
@@ -1,6 +1,6 @@
-import { chartConfig } from '../../config';
-import { isFiniteNumber } from '../../utils/numbers';
-import type { ChartInset, Series } from '../../utils/types';
+import { chartConfig } from '../config';
+import { isFiniteNumber } from '../utils/numbers';
+import type { ChartInset, Series } from '../utils/types';
 import { DEFAULT_AXIS_HEIGHT } from '../Axis/XAxis';
 import type { XAxisProps } from '../Axis/XAxis';
 import { DEFAULT_AXIS_WIDTH } from '../Axis/YAxis';
