+++ libs/ui-rnative/src/lib/Components/visualization/Point/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
 
-import { chartConfig } from '../../config';
+import { chartConfig } from '../config';
 import {
   buildArrowPoints,
   computeLabelX,
