+++ libs/ui-rnative/src/lib/Components/visualization/Scrubber/DefaultScrubberTooltip/ChartTooltipItem.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { G, Text as SvgText } from 'react-native-svg';
 
-import { chartConfig, useChartTokens } from '../../../config';
+import { chartConfig, useChartTokens } from '../../config';
 import type { ChartTooltipItemProps } from '../types';
 
 /**
