+++ libs/ui-rnative/src/lib/Components/visualization/Scrubber/ScrubberProvider.tsx
@@ -5,8 +5,8 @@ import { Gesture, GestureDetector } from 'react-native-gesture-handler';
 import Animated from 'react-native-reanimated';
 import { scheduleOnRN } from 'react-native-worklets';
 
-import { chartConfig } from '../../config';
-import { clamp } from '../../utils/numbers';
+import { chartConfig } from '../config';
+import { clamp } from '../utils/numbers';
 import { useCartesianChartContext } from '../CartesianChart/context';
 import { useMagneticSnapshot } from '../Point/pointContext';
 import { ScrubberContextProvider } from './context';
