+++ libs/ui-rnative/src/lib/Components/visualization/utils/scales/scales.ts
@@ -1,6 +1,6 @@
 import { scaleBand, scaleLinear, scaleLog } from 'd3-scale';
 
-import type { AxisBounds, BaseAxisProps } from '../../Components/Axis';
+import type { AxisBounds, BaseAxisProps } from '../../Axis';
 import type {
   CategoricalScale,
   ChartScaleFunction,
